revision de codigo

+++ src/main/java/com/cris/bank_app_backend/entities/PrestamoEntity.java
@@ -21,14 +21,14 @@ public class PrestamoEntity {
     private String tipoPrestamo;
     private int plazo;
     private int numeroDeCuotas;
-    private int montoDelPrestamo;
+    private double montoDelPrestamo;
     private double tasaDeInteresAnual;
     private double tasaDeInteresMensual;
-    private int cuotaMensual;
-    private int seguroDeDesgravamen;
-    private int seguroDeIncendio;
-    private int comisionPorAdministracion;
-    private int costoMensual;
-    private int costosTotales;
+    private double cuotaMensual;
+    private double seguroDeDesgravamen;
+    private double seguroDeIncendio;
+    private double comisionPorAdministracion;
+    private double costoMensual;
+    private double costosTotales;
     private Long solicitudId;
 }
